drone_plugins/healthchecker/requirements.txt
ansicolors==1.1.8
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
mypy==1.8.0
mypy-extensions==1.0.0
requests==2.31.0
types-requests==2.31.0.10
typing_extensions==4.9.0
urllib3==2.1.0
